add maxfw
