+++ .dockerignore
@@ -1,5 +1,4 @@
 *.md
-Dockerfile
 .git/
 .idea/
 .pytest_cache/
