+++ lm_1b/lm_1b_eval.py
@@ -64,7 +64,6 @@ tf.flags.DEFINE_string('input_data', '',
 tf.flags.DEFINE_integer('max_eval_steps', 1000000,
                         'Maximum mumber of steps to run "eval" mode.')
 
-
 # For saving demo resources, use batch size 1 and step 1.
 BATCH_SIZE = 1
 NUM_TIMESTEPS = 1
@@ -72,240 +71,239 @@ MAX_WORD_LEN = 50
 
 
 def _LoadModel(gd_file, ckpt_file):
-  """Load the model from GraphDef and Checkpoint.
+    """Load the model from GraphDef and Checkpoint.
 
-  Args:
-    gd_file: GraphDef proto text file.
-    ckpt_file: TensorFlow Checkpoint file.
+    Args:
+      gd_file: GraphDef proto text file.
+      ckpt_file: TensorFlow Checkpoint file.
 
-  Returns:
-    TensorFlow session and tensors dict.
-  """
-  with tf.Graph().as_default():
-    sys.stderr.write('Recovering graph.\n')
-    with tf.gfile.FastGFile(gd_file, 'r') as f:
-      s = f.read() #.decode()
-      gd = tf.GraphDef()
-      text_format.Merge(s, gd)
+    Returns:
+      TensorFlow session and tensors dict.
+    """
+    with tf.Graph().as_default():
+        sys.stderr.write('Recovering graph.\n')
+        with tf.gfile.FastGFile(gd_file, 'r') as f:
+            s = f.read()  # .decode()
+            gd = tf.GraphDef()
+            text_format.Merge(s, gd)
 
-    tf.logging.info('Recovering Graph %s', gd_file)
-    t = {}
-    [t['states_init'], t['lstm/lstm_0/control_dependency'],
-     t['lstm/lstm_1/control_dependency'], t['softmax_out'], t['class_ids_out'],
-     t['class_weights_out'], t['log_perplexity_out'], t['inputs_in'],
-     t['targets_in'], t['target_weights_in'], t['char_inputs_in'],
-     t['all_embs'], t['softmax_weights'], t['global_step']
-    ] = tf.import_graph_def(gd, {}, ['states_init',
-                                     'lstm/lstm_0/control_dependency:0',
-                                     'lstm/lstm_1/control_dependency:0',
-                                     'softmax_out:0',
-                                     'class_ids_out:0',
-                                     'class_weights_out:0',
-                                     'log_perplexity_out:0',
-                                     'inputs_in:0',
-                                     'targets_in:0',
-                                     'target_weights_in:0',
-                                     'char_inputs_in:0',
-                                     'all_embs_out:0',
-                                     'Reshape_3:0',
-                                     'global_step:0'], name='')
+        tf.logging.info('Recovering Graph %s', gd_file)
+        t = {}
+        [t['states_init'], t['lstm/lstm_0/control_dependency'],
+         t['lstm/lstm_1/control_dependency'], t['softmax_out'], t['class_ids_out'],
+         t['class_weights_out'], t['log_perplexity_out'], t['inputs_in'],
+         t['targets_in'], t['target_weights_in'], t['char_inputs_in'],
+         t['all_embs'], t['softmax_weights'], t['global_step']
+         ] = tf.import_graph_def(gd, {}, ['states_init',
+                                          'lstm/lstm_0/control_dependency:0',
+                                          'lstm/lstm_1/control_dependency:0',
+                                          'softmax_out:0',
+                                          'class_ids_out:0',
+                                          'class_weights_out:0',
+                                          'log_perplexity_out:0',
+                                          'inputs_in:0',
+                                          'targets_in:0',
+                                          'target_weights_in:0',
+                                          'char_inputs_in:0',
+                                          'all_embs_out:0',
+                                          'Reshape_3:0',
+                                          'global_step:0'], name='')
 
-    sys.stderr.write('Recovering checkpoint %s\n' % ckpt_file)
-    sess = tf.Session(config=tf.ConfigProto(allow_soft_placement=True))
-    sess.run('save/restore_all', {'save/Const:0': ckpt_file})
-    sess.run(t['states_init'])
+        sys.stderr.write('Recovering checkpoint %s\n' % ckpt_file)
+        sess = tf.Session(config=tf.ConfigProto(allow_soft_placement=True))
+        sess.run('save/restore_all', {'save/Const:0': ckpt_file})
+        sess.run(t['states_init'])
 
-  return sess, t
+    return sess, t
 
 
 def _EvalModel(dataset):
-  """Evaluate model perplexity using provided dataset.
+    """Evaluate model perplexity using provided dataset.
 
-  Args:
-    dataset: LM1BDataset object.
-  """
-  sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
+    Args:
+      dataset: LM1BDataset object.
+    """
+    sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
 
-  current_step = t['global_step'].eval(session=sess)
-  sys.stderr.write('Loaded step %d.\n' % current_step)
+    current_step = t['global_step'].eval(session=sess)
+    sys.stderr.write('Loaded step %d.\n' % current_step)
 
-  data_gen = dataset.get_batch(BATCH_SIZE, NUM_TIMESTEPS, forever=False)
-  sum_num = 0.0
-  sum_den = 0.0
-  perplexity = 0.0
-  for i, (inputs, char_inputs, _, targets, weights) in enumerate(data_gen):
-    input_dict = {t['inputs_in']: inputs,
-                  t['targets_in']: targets,
-                  t['target_weights_in']: weights}
-    if 'char_inputs_in' in t:
-      input_dict[t['char_inputs_in']] = char_inputs
-    log_perp = sess.run(t['log_perplexity_out'], feed_dict=input_dict)
+    data_gen = dataset.get_batch(BATCH_SIZE, NUM_TIMESTEPS, forever=False)
+    sum_num = 0.0
+    sum_den = 0.0
+    perplexity = 0.0
+    for i, (inputs, char_inputs, _, targets, weights) in enumerate(data_gen):
+        input_dict = {t['inputs_in']: inputs,
+                      t['targets_in']: targets,
+                      t['target_weights_in']: weights}
+        if 'char_inputs_in' in t:
+            input_dict[t['char_inputs_in']] = char_inputs
+        log_perp = sess.run(t['log_perplexity_out'], feed_dict=input_dict)
 
-    if np.isnan(log_perp):
-      sys.stderr.error('log_perplexity is Nan.\n')
-    else:
-      sum_num += log_perp * weights.mean()
-      sum_den += weights.mean()
-    if sum_den > 0:
-      perplexity = np.exp(sum_num / sum_den)
+        if np.isnan(log_perp):
+            sys.stderr.error('log_perplexity is Nan.\n')
+        else:
+            sum_num += log_perp * weights.mean()
+            sum_den += weights.mean()
+        if sum_den > 0:
+            perplexity = np.exp(sum_num / sum_den)
 
-    sys.stderr.write('Eval Step: %d, Average Perplexity: %f.\n' %
-                     (i, perplexity))
+        sys.stderr.write('Eval Step: %d, Average Perplexity: %f.\n' %
+                         (i, perplexity))
 
-    if i > FLAGS.max_eval_steps:
-      break
+        if i > FLAGS.max_eval_steps:
+            break
 
 
 def _SampleSoftmax(softmax):
-  return min(np.sum(np.cumsum(softmax) < np.random.rand()), len(softmax) - 1)
+    return min(np.sum(np.cumsum(softmax) < np.random.rand()), len(softmax) - 1)
 
 
 def _SampleModel(prefix_words, vocab):
-  """Predict next words using the given prefix words.
+    """Predict next words using the given prefix words.
 
-  Args:
-    prefix_words: Prefix words.
-    vocab: Vocabulary. Contains max word chard id length and converts between
-        words and ids.
-  """
-  targets = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
-  weights = np.ones([BATCH_SIZE, NUM_TIMESTEPS], np.float32)
+    Args:
+      prefix_words: Prefix words.
+      vocab: Vocabulary. Contains max word chard id length and converts between
+          words and ids.
+    """
+    targets = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
+    weights = np.ones([BATCH_SIZE, NUM_TIMESTEPS], np.float32)
 
-  sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
+    sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
 
-  if prefix_words.find('<S>') != 0:
-    prefix_words = '<S> ' + prefix_words
+    if prefix_words.find('<S>') != 0:
+        prefix_words = '<S> ' + prefix_words
 
-  prefix = [vocab.word_to_id(w) for w in prefix_words.split()]
-  prefix_char_ids = [vocab.word_to_char_ids(w) for w in prefix_words.split()]
-  for _ in xrange(FLAGS.num_samples):
-    inputs = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
-    char_ids_inputs = np.zeros(
-        [BATCH_SIZE, NUM_TIMESTEPS, vocab.max_word_length], np.int32)
-    samples = prefix[:]
-    char_ids_samples = prefix_char_ids[:]
-    sent = ''
-    while True:
-      inputs[0, 0] = samples[0]
-      char_ids_inputs[0, 0, :] = char_ids_samples[0]
-      samples = samples[1:]
-      char_ids_samples = char_ids_samples[1:]
+    prefix = [vocab.word_to_id(w) for w in prefix_words.split()]
+    prefix_char_ids = [vocab.word_to_char_ids(w) for w in prefix_words.split()]
+    for _ in xrange(FLAGS.num_samples):
+        inputs = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
+        char_ids_inputs = np.zeros(
+            [BATCH_SIZE, NUM_TIMESTEPS, vocab.max_word_length], np.int32)
+        samples = prefix[:]
+        char_ids_samples = prefix_char_ids[:]
+        sent = ''
+        while True:
+            inputs[0, 0] = samples[0]
+            char_ids_inputs[0, 0, :] = char_ids_samples[0]
+            samples = samples[1:]
+            char_ids_samples = char_ids_samples[1:]
 
-      softmax = sess.run(t['softmax_out'],
-                         feed_dict={t['char_inputs_in']: char_ids_inputs,
-                                    t['inputs_in']: inputs,
-                                    t['targets_in']: targets,
-                                    t['target_weights_in']: weights})
+            softmax = sess.run(t['softmax_out'],
+                               feed_dict={t['char_inputs_in']: char_ids_inputs,
+                                          t['inputs_in']: inputs,
+                                          t['targets_in']: targets,
+                                          t['target_weights_in']: weights})
 
-      sample = _SampleSoftmax(softmax[0])
-      sample_char_ids = vocab.word_to_char_ids(vocab.id_to_word(sample))
+            sample = _SampleSoftmax(softmax[0])
+            sample_char_ids = vocab.word_to_char_ids(vocab.id_to_word(sample))
 
-      if not samples:
-        samples = [sample]
-        char_ids_samples = [sample_char_ids]
-      sent += vocab.id_to_word(samples[0]) + ' '
-      sys.stderr.write('%s\n' % sent)
+            if not samples:
+                samples = [sample]
+                char_ids_samples = [sample_char_ids]
+            sent += vocab.id_to_word(samples[0]) + ' '
+            sys.stderr.write('%s\n' % sent)
 
-      if (vocab.id_to_word(samples[0]) == '</S>' or len(sent) > FLAGS.max_sample_words):
-          txt_file = open("out.txt", "w")
-          txt_file.write(str(sent))
-          txt_file.close()
-          break
-
+            if (vocab.id_to_word(samples[0]) == '</S>' or len(sent) > FLAGS.max_sample_words):
+                txt_file = open("out.txt", "w")
+                txt_file.write(str(sent))
+                txt_file.close()
+                break
 
 
 def _DumpEmb(vocab):
-  """Dump the softmax weights and word embeddings to files.
+    """Dump the softmax weights and word embeddings to files.
 
-  Args:
-    vocab: Vocabulary. Contains vocabulary size and converts word to ids.
-  """
-  assert FLAGS.save_dir, 'Must specify FLAGS.save_dir for dump_emb.'
-  inputs = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
-  targets = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
-  weights = np.ones([BATCH_SIZE, NUM_TIMESTEPS], np.float32)
+    Args:
+      vocab: Vocabulary. Contains vocabulary size and converts word to ids.
+    """
+    assert FLAGS.save_dir, 'Must specify FLAGS.save_dir for dump_emb.'
+    inputs = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
+    targets = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
+    weights = np.ones([BATCH_SIZE, NUM_TIMESTEPS], np.float32)
 
-  sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
+    sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
 
-  softmax_weights = sess.run(t['softmax_weights'])
-  fname = FLAGS.save_dir + '/embeddings_softmax.npy'
-  with tf.gfile.Open(fname, mode='w') as f:
-    np.save(f, softmax_weights)
-  sys.stderr.write('Finished softmax weights\n')
+    softmax_weights = sess.run(t['softmax_weights'])
+    fname = FLAGS.save_dir + '/embeddings_softmax.npy'
+    with tf.gfile.Open(fname, mode='w') as f:
+        np.save(f, softmax_weights)
+    sys.stderr.write('Finished softmax weights\n')
 
-  all_embs = np.zeros([vocab.size, 1024])
-  for i in xrange(vocab.size):
-    input_dict = {t['inputs_in']: inputs,
-                  t['targets_in']: targets,
-                  t['target_weights_in']: weights}
-    if 'char_inputs_in' in t:
-      input_dict[t['char_inputs_in']] = (
-          vocab.word_char_ids[i].reshape([-1, 1, MAX_WORD_LEN]))
-    embs = sess.run(t['all_embs'], input_dict)
-    all_embs[i, :] = embs
-    sys.stderr.write('Finished word embedding %d/%d\n' % (i, vocab.size))
+    all_embs = np.zeros([vocab.size, 1024])
+    for i in xrange(vocab.size):
+        input_dict = {t['inputs_in']: inputs,
+                      t['targets_in']: targets,
+                      t['target_weights_in']: weights}
+        if 'char_inputs_in' in t:
+            input_dict[t['char_inputs_in']] = (
+                vocab.word_char_ids[i].reshape([-1, 1, MAX_WORD_LEN]))
+        embs = sess.run(t['all_embs'], input_dict)
+        all_embs[i, :] = embs
+        sys.stderr.write('Finished word embedding %d/%d\n' % (i, vocab.size))
 
-  fname = FLAGS.save_dir + '/embeddings_char_cnn.npy'
-  with tf.gfile.Open(fname, mode='w') as f:
-    np.save(f, all_embs)
-  sys.stderr.write('Embedding file saved\n')
+    fname = FLAGS.save_dir + '/embeddings_char_cnn.npy'
+    with tf.gfile.Open(fname, mode='w') as f:
+        np.save(f, all_embs)
+    sys.stderr.write('Embedding file saved\n')
 
 
 def _DumpSentenceEmbedding(sentence, vocab):
-  """Predict next words using the given prefix words.
+    """Predict next words using the given prefix words.
 
-  Args:
-    sentence: Sentence words.
-    vocab: Vocabulary. Contains max word chard id length and converts between
-        words and ids.
-  """
-  targets = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
-  weights = np.ones([BATCH_SIZE, NUM_TIMESTEPS], np.float32)
+    Args:
+      sentence: Sentence words.
+      vocab: Vocabulary. Contains max word chard id length and converts between
+          words and ids.
+    """
+    targets = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
+    weights = np.ones([BATCH_SIZE, NUM_TIMESTEPS], np.float32)
 
-  sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
+    sess, t = _LoadModel(FLAGS.pbtxt, FLAGS.ckpt)
 
-  if sentence.find('<S>') != 0:
-    sentence = '<S> ' + sentence
+    if sentence.find('<S>') != 0:
+        sentence = '<S> ' + sentence
 
-  word_ids = [vocab.word_to_id(w) for w in sentence.split()]
-  char_ids = [vocab.word_to_char_ids(w) for w in sentence.split()]
+    word_ids = [vocab.word_to_id(w) for w in sentence.split()]
+    char_ids = [vocab.word_to_char_ids(w) for w in sentence.split()]
 
-  inputs = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
-  char_ids_inputs = np.zeros(
-      [BATCH_SIZE, NUM_TIMESTEPS, vocab.max_word_length], np.int32)
-  for i in xrange(len(word_ids)):
-    inputs[0, 0] = word_ids[i]
-    char_ids_inputs[0, 0, :] = char_ids[i]
+    inputs = np.zeros([BATCH_SIZE, NUM_TIMESTEPS], np.int32)
+    char_ids_inputs = np.zeros(
+        [BATCH_SIZE, NUM_TIMESTEPS, vocab.max_word_length], np.int32)
+    for i in xrange(len(word_ids)):
+        inputs[0, 0] = word_ids[i]
+        char_ids_inputs[0, 0, :] = char_ids[i]
 
-    # Add 'lstm/lstm_0/control_dependency' if you want to dump previous layer
-    # LSTM.
-    lstm_emb = sess.run(t['lstm/lstm_1/control_dependency'],
-                        feed_dict={t['char_inputs_in']: char_ids_inputs,
-                                   t['inputs_in']: inputs,
-                                   t['targets_in']: targets,
-                                   t['target_weights_in']: weights})
+        # Add 'lstm/lstm_0/control_dependency' if you want to dump previous layer
+        # LSTM.
+        lstm_emb = sess.run(t['lstm/lstm_1/control_dependency'],
+                            feed_dict={t['char_inputs_in']: char_ids_inputs,
+                                       t['inputs_in']: inputs,
+                                       t['targets_in']: targets,
+                                       t['target_weights_in']: weights})
 
-    fname = os.path.join(FLAGS.save_dir, 'lstm_emb_step_%d.npy' % i)
-    with tf.gfile.Open(fname, mode='w') as f:
-      np.save(f, lstm_emb)
-    sys.stderr.write('LSTM embedding step %d file saved\n' % i)
+        fname = os.path.join(FLAGS.save_dir, 'lstm_emb_step_%d.npy' % i)
+        with tf.gfile.Open(fname, mode='w') as f:
+            np.save(f, lstm_emb)
+        sys.stderr.write('LSTM embedding step %d file saved\n' % i)
 
 
 def main(unused_argv):
-  vocab = data_utils.CharsVocabulary(FLAGS.vocab_file, MAX_WORD_LEN)
+    vocab = data_utils.CharsVocabulary(FLAGS.vocab_file, MAX_WORD_LEN)
 
-  if FLAGS.mode == 'eval':
-    dataset = data_utils.LM1BDataset(FLAGS.input_data, vocab)
-    _EvalModel(dataset)
-  elif FLAGS.mode == 'sample':
-    _SampleModel(FLAGS.prefix, vocab)
-  elif FLAGS.mode == 'dump_emb':
-    _DumpEmb(vocab)
-  elif FLAGS.mode == 'dump_lstm_emb':
-    _DumpSentenceEmbedding(FLAGS.sentence, vocab)
-  else:
-    raise Exception('Mode not supported.')
+    if FLAGS.mode == 'eval':
+        dataset = data_utils.LM1BDataset(FLAGS.input_data, vocab)
+        _EvalModel(dataset)
+    elif FLAGS.mode == 'sample':
+        _SampleModel(FLAGS.prefix, vocab)
+    elif FLAGS.mode == 'dump_emb':
+        _DumpEmb(vocab)
+    elif FLAGS.mode == 'dump_lstm_emb':
+        _DumpSentenceEmbedding(FLAGS.sentence, vocab)
+    else:
+        raise Exception('Mode not supported.')
 
 
 if __name__ == '__main__':
-  tf.app.run()
+    tf.app.run()
